decomposed Navbar into child components: MenuDrawer for the mobile menu drawer, and MainNav for the Main Navbar. Will Do the same for the Checkout Drawer next.
